refactor: wording
