+++ packages/workflow/src/config/routes.ts
@@ -29,7 +29,7 @@ import { approveCorrectionRoute } from '@workflow/records/handler/correction/app
 import { requestCorrectionRoute } from '@workflow/records/handler/correction/request'
 import { makeCorrectionRoute } from '@workflow/records/handler/correction/make-correction'
 import { eventNotificationHandler } from '@workflow/records/handler/eventNotificationHandler'
-import { upsertRegistrationHandler } from '@workflow/records/handler/upsert'
+import { upsertRegistrationHandler } from '@workflow/records/handler/upsert-identifiers'
 
 export const getRoutes = () => {
   const routes = [
@@ -72,7 +72,7 @@ export const getRoutes = () => {
     },
     {
       method: 'POST',
-      path: '/records/{id}/upsert',
+      path: '/records/{id}/upsert-identifiers',
       handler: upsertRegistrationHandler,
       config: {
         tags: ['api'],
